+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## LibFrameFade
+### v4
+- Reimplement UIFrameIsFading to return correct results for addons that need it.
+  - This causes the UIFrameIsFading to always taint if called, however the function is unused by Blizzard in all current (9.1.0, 2.5.2, and 1.14.0) versions of the game.
+
 ### v3
 - Fixed an issue where ElvUI's chat frame tabs would incorrectly fade out when not actively hovered over.